Add grid layout to anchors in aside element
To apply the effect of aligning and spacing of nav links, the grid
layout and properties were applied to the anchor elements that reside
in the aside element.

For this to work, the elements that were parents of the anchor elements
(ul, li) had their layout changed to contents, so that inheriting the
defined grid values (specifically grid-template-columns)
from the aside element to the anchor elements were streamlined.