Change cursor icon on hover for empty nav links
With how the nav list in the aside element is structured, there are
empty nav links purely for creating big gaps between certain links.
However, because they are still anchor elements, hovering over them
still changes the cursor to a 'pointer', indicating to the user that
it is clickable.

For those empty nav links, the cursor was changed to the default icon,
to remove the impression that it is a working link.