Add grid layout to ul element in .announcements
This is to center the list items in the ul within the .announcements
div, as well as provide spacing between list items.

Additionally, the widths of the list item's have been reduced. This is
not only to follow the given design, but also to fix the issue of the
hr element shrinking to a dot size when using the 'justify-items'
property with value of 'center'.