Add grid layout to li element in .trending
This is to complete the final part of the layout, where the profile pic
sits in the first column, and the user's name and their trending project
in the second column. This is done with the 'grid-column' grid item
property, so that the list item can take up two columns, and properly
assign the items (img and text elements) in their respective columns.